email can now duplicate
this is okay because username is used to decide user uniqueness. If a user has both a professor and admin account for example, it is acceptable.

It is also so that admin can upsert changes via csv easily. we can change this to have table dependent endpoints once the table values are set in stone.