Db migrations (#85)
* migrations file runs

created and enabled migrations file. now to make it actually get migrations

* now always runs all migrations

next to make it check if migrations are present. this means making a table to store present ones and checking length of table with length of files

* migrations now run

if migrations table doesn't exist, it will display an error and run the migrations migration to enable migrations :)

* Update runMigrations.js

* Update runMigrations.js

* modules working as intended

await migrations now properly awaits on top level

* finish module migration

* adjust db query

This means that the ID of the user must be in the instructor table to count as an instructor when importing sections

* Update server/api/admin.js

Co-authored-by: Jonatan Schroeder <jonatan@yorku.ca>

---------

Co-authored-by: Jonatan Schroeder <jonatan@yorku.ca>