Dev (#115)
* Update README.md

* abstract axios function

* make it send to 403

* begin replace axios with wrapped

* add 403 page and route

* replace axios in dashboard

* replace axios with wrapped get

* remove axios

* Update professor_dash.js

* Update profile_view.js

* Update header.js

* Update student_dash.js

* Update login.js

* Update profile.js

removed big unused code

* remove unused

* Update student_profile.js

* Update professor_section.js

* add break for 500

so the user doesn't get navigated on server error (ie. invalid form data)

* Update requestWrapper.js

* change from nav to reload

* wrapper now throws error back to parent

* only throws err on non-403 error

this is because it handles 403 by refreshing the page

* remove 500 case

* add error alert component

* alert should display on post failure

only in student application

* Update requestWrapper.js

* add pubsub and audit fix force

* add alert using pubsub

* found fix for weird hover artifacts

* fix warning

* make rating revert on failed post

* Remove unused component

ErrorAlert is now not used anywhere, but it could be handy so I will leave it in as a component to maybe use later.

* fix key requirement for list items

* solve x cannot be child of y warning

resolved by adding a tag that makes it so the parent node is a <div> instead of <p>
MUI issue found here: https://github.com/mui/material-ui/issues/32289

* fix nth-of-type warning

* move files

* remove comment

* add first migration

* Update index.js

* Db migrations (#85)

* migrations file runs

created and enabled migrations file. now to make it actually get migrations

* now always runs all migrations

next to make it check if migrations are present. this means making a table to store present ones and checking length of table with length of files

* migrations now run

if migrations table doesn't exist, it will display an error and run the migrations migration to enable migrations :)

* Update runMigrations.js

* Update runMigrations.js

* modules working as intended

await migrations now properly awaits on top level

* finish module migration

* adjust db query

This means that the ID of the user must be in the instructor table to count as an instructor when importing sections

* Update server/api/admin.js

Co-authored-by: Jonatan Schroeder <jonatan@yorku.ca>

---------

Co-authored-by: Jonatan Schroeder <jonatan@yorku.ca>

* Term wizard (#86)

* term wizard page added

no functionality yet

* prettify

* create term endpoint

* add table

* table data populates properly

* add default sort

* Update datagrid_table.js

* Add endpoint for updating term

* Rename and add post feature

Currently has CRU out of CRUD, do we want delete capabilities? Or rename? I don't think it's important but maybe.

* npm update (again server)

* App push on apply (#97)

* minor refactor + FE support

* allow nav to be null

just in case it needs to be from a use source

* backend properly prints latest courses

next to make it insert to DB

* frontend properly posts

* clicking "apply" now pulls app

NOTE: doesn't work if navigating to the term via URL!!!

* formatting

* todo

* FE queries extra endpoint

* abstract endpoint

also add filtering and everything should work now

* remove on click

doubled effect, now runs when accessing via URL

* fix 2 bugs

- endpoint not sending response
- FE sending null to endpoint
security issue remains: if null is sent to endpoint, server might crash

* 404 page now properly shows when term doesn't exist

* remove unused var

* redo func

* npm install

* npm updates

* Endpoint error handler (#113)

* add error catching

* add package

* add handler to index

* add handler to endpoints

* add handler to applicant

* add handler to auth

* Update instructor.js

* change AS to its own import everywhere

---------

Co-authored-by: kukuminer <liranzh2000@gmail.com>
Co-authored-by: kukuminer <36781630+kukuminer@users.noreply.github.com>